+++ _env.crash
@@ -1,0 +1,2 @@
+# REACT_APP_API_URL=http://localhost:3000/api
+REACT_APP_API_URL=https://soyummy-qk5m.onrender.com/api
